create kernel module;move non-cats-effectful members to kernel module;
