+++ kernel/src/main/scala/tofu/control/impl/FunctorDelegate.scala
