+++ kernel/src/main/scala/tofu/internal/FoldableStream.scala
