+++ kernel/src/main/scala/tofu/internal/Newtype1Covariant.scala
