+++ kernel/src/main/scala/tofu/syntax/feither.scala
