+++ kernel/src/main/scala/tofu/syntax/foldable.scala
